// src/api/api.ts
export const BASE_PATH = 'http://peterserver1.myvnc.com:8080/rest'.replace(/\/+$/, '');

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
  csv: ',',
  ssv: ' ',
  tsv: '\t',
  pipes: '|'
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
  (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
  url: string;
  options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
  protected configuration: Configuration = new Configuration();

  constructor(
    configuration?: Configuration,
    protected basePath: string = BASE_PATH,
    protected fetch: FetchAPI = portableFetch
  ) {
    if (configuration) {
      this.configuration = configuration;
      this.basePath = configuration.basePath || this.basePath;
    }
  }
}
// ...
export interface CommandDescription {
  /**
   *
   * @type {Array<CommandOption>}
   * @memberof CommandDescription
   */
  commandOptions?: Array<CommandOption>;
}
// ...
export interface ConfigurableServiceDTO {
  /**
   *
   * @type {string}
   * @memberof ConfigurableServiceDTO
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof ConfigurableServiceDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof ConfigurableServiceDTO
   */
  category?: string;
  /**
   *
   * @type {string}
   * @memberof ConfigurableServiceDTO
   */
  configDescriptionURI?: string;
  /**
   *
   * @type {boolean}
   * @memberof ConfigurableServiceDTO
   */
  multiple?: boolean;
}
// ...
export interface EnrichedItemDTO {
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  category?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  tags?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  groupNames?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  link?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  state?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  transformedState?: string;
  /**
   *
   * @type {StateDescription}
   * @memberof EnrichedItemDTO
   */
  stateDescription?: StateDescription;
  /**
   *
   * @type {CommandDescription}
   * @memberof EnrichedItemDTO
   */
  commandDescription?: CommandDescription;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof EnrichedItemDTO
   */
  metadata?: { [key: string]: any };
  /**
   *
   * @type {boolean}
   * @memberof EnrichedItemDTO
   */
  editable?: boolean;
}
// ...
export interface MappingDTO {
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  command?: string;
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  label?: string;
}

/**
 *
 * @export
 * @interface MetadataDTO
 */
export interface MetadataDTO {
  /**
   *
   * @type {string}
   * @memberof MetadataDTO
   */
  value?: string;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof MetadataDTO
   */
  config?: { [key: string]: any };
}

/**
 *
 * @export
 * @interface PageDTO
 */
export interface PageDTO {
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  title?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  link?: string;
  /**
   *
   * @type {PageDTO}
   * @memberof PageDTO
   */
  parent?: PageDTO;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  leaf?: boolean;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  timeout?: boolean;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof PageDTO
   */
  widgets?: Array<WidgetDTO>;
}
// ...
export interface StateDescription {
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  minimum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  maximum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof StateDescription
   */
  pattern?: string;
  /**
   *
   * @type {boolean}
   * @memberof StateDescription
   */
  readOnly?: boolean;
  /**
   *
   * @type {Array<StateOption>}
   * @memberof StateDescription
   */
  options?: Array<StateOption>;
}
// ...
export interface WidgetDTO {
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  widgetId?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  name?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  visibility?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  labelcolor?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  valuecolor?: string;
  /**
   *
   * @type {Array<MappingDTO>}
   * @memberof WidgetDTO
   */
  mappings?: Array<MappingDTO>;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  switchSupport?: boolean;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  sendFrequency?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  separator?: string;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  refresh?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  height?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  minValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  maxValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  url?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  encoding?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  service?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  period?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  legend?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  state?: string;
  /**
   *
   * @type {EnrichedItemDTO}
   * @memberof WidgetDTO
   */
  item?: EnrichedItemDTO;
  /**
   *
   * @type {PageDTO}
   * @memberof WidgetDTO
   */
  linkedPage?: PageDTO;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof WidgetDTO
   */
  widgets?: Array<WidgetDTO>;
}
// ...
export interface ServicesApiInterface {
  /**
   *
   * @summary Deletes a service configuration for given service ID and returns the old configuration.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  deleteConfiguration(serviceId: string, options?: any): Promise<string>;

  /**
   *
   * @summary Get all configurable services.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  getAll(options?: any): Promise<Array<ConfigurableServiceDTO>>;

  /**
   *
   * @summary Get configurable service for given service ID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  getById(serviceId: string, options?: any): Promise<ConfigurableServiceDTO>;

  /**
   *
   * @summary Get service configuration for given service ID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  getConfiguration(serviceId: string, options?: any): Promise<string>;

  /**
   *
   * @summary Get existing multiple context service configurations for the given factory PID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  getMultiConfigServicesByFactoryPid(
    serviceId: string,
    options?: any
  ): Promise<Array<ConfigurableServiceDTO>>;

  /**
   *
   * @summary Updates a service configuration for given service ID and returns the old configuration.
   * @param {string} serviceId service ID
   * @param {any} [body]
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApiInterface
   */
  updateConfiguration(serviceId: string, body?: any, options?: any): Promise<string>;
}

/**
 * ServicesApi - object-oriented interface
 * @export
 * @class ServicesApi
 * @extends {BaseAPI}
 */
export class ServicesApi extends BaseAPI implements ServicesApiInterface {
  /**
   *
   * @summary Deletes a service configuration for given service ID and returns the old configuration.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public deleteConfiguration(serviceId: string, options?: any) {
    return ServicesApiFp(this.configuration).deleteConfiguration(serviceId, options)(
      this.fetch,
      this.basePath
    );
  }

  /**
   *
   * @summary Get all configurable services.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public getAll(options?: any) {
    return ServicesApiFp(this.configuration).getAll(options)(this.fetch, this.basePath);
  }

  /**
   *
   * @summary Get configurable service for given service ID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public getById(serviceId: string, options?: any) {
    return ServicesApiFp(this.configuration).getById(serviceId, options)(this.fetch, this.basePath);
  }

  /**
   *
   * @summary Get service configuration for given service ID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public getConfiguration(serviceId: string, options?: any) {
    return ServicesApiFp(this.configuration).getConfiguration(serviceId, options)(
      this.fetch,
      this.basePath
    );
  }

  /**
   *
   * @summary Get existing multiple context service configurations for the given factory PID.
   * @param {string} serviceId service ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public getMultiConfigServicesByFactoryPid(serviceId: string, options?: any) {
    return ServicesApiFp(this.configuration).getMultiConfigServicesByFactoryPid(serviceId, options)(
      this.fetch,
      this.basePath
    );
  }

  /**
   *
   * @summary Updates a service configuration for given service ID and returns the old configuration.
   * @param {string} serviceId service ID
   * @param {any} [body]
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ServicesApi
   */
  public updateConfiguration(serviceId: string, body?: any, options?: any) {
    return ServicesApiFp(this.configuration).updateConfiguration(
      serviceId,
      body,
      options
    )(this.fetch, this.basePath);
  }
}